Move GeoPolyActivityTest to geo module
