+++ collect_app/src/main/java/org/odk/collect/android/application/Collect.java
@@ -96,7 +96,7 @@ public class Collect extends Application implements
 
     private AudioRecorderDependencyComponent audioRecorderDependencyComponent;
     private ProjectsDependencyComponent projectsDependencyComponent;
-    public GeoDependencyComponent geoDependencyComponent;
+    private GeoDependencyComponent geoDependencyComponent;
 
     /**
      * @deprecated we shouldn't have to reference a static singleton of the application. Code doing this
