+++ geo/src/main/res/drawable/ic_backspace.xml
