+++ geo/src/main/res/drawable/ic_pause_36.xml
